Fix remapped diagnostics
For column integrals of remapped outputs to conserve, the layer thicknesses
that the diagnostic manager should use might be different from the one
associated with the last time diag_cs%h was updated. For example, the sum
of uhtr and vhtr should be associated with the thicknesses before the
dynamics step because h_end = h_pre + div(uhtr) + div(vhtr). This is part
of a larger effort with @adcroft to really figure out how to ensure all
diagnostics are properly posted when they are remapped.