Enforce conservation of the remapped quantity
In the case that the source and destination columns are the same thickness,
the column integrals of the remapped quantities should be the same. To do
this, the difference between the column integrals from the source and
destination is added to the cell in the destination with the largest
amount.