Fixes problems with the updating of offline fields
Both updating from files and updating from arrays had errors. When
updating from files, an entire 2D slice of the array was being filled with
a single file for all levels greater than nk_input. Updating from arrays
was filling temperature and salinity with 0 instead of flooding the cells
with the lowest valid level of the input fields. Kd is now also flooded
when read in and does not need to be flooded again in offline_diabatic_ale.