Change diagnostic reintegration using fraction of total layer
The original algorithm uses the fraction of the remaining thickness of the
source cell to calculate the fractional amount of quantity to add to the
destination column. This potentially leads to destination cells, spanning
the same source cell, with the same thickness having different amounts of
the quantity because of roundoff. This commit calculates a fraction based
on the entire thickness of the source cell which seems to make the
reintegration more accurate.