Add data streaming functionality
Initialization routines and hooks into the code have been added to
enable the streaming of data out of the model to an appropriate
client that can receive 2D or 3D 64-bit arrays. Diagnostics to be
sent are specified via the diag_table. Pointers to the requested arrays
are set at initialization for diagnostics requested via the diag_table.