Add uniform z-like grid to OPT BC grid
To avoid layers which are ~1000m thick near the bottom of the ocean in
the unstratified abysal/deep ocean, this adds the calculation of a
a uniform grid that spans the column between surface boundary layer
depth and the bottom of the ocean. This z-like grid is then merged with
the Gauss-Lobato grid to form the final grid that the state is remapped to.