Incorporate Scott's reimplementation of fixed z*
Scott rejiggered a previous iteration of the code where we
independently create a z* like grid with a Gauss-Lobato grid.
This iteration determines the number of GL points to be used
based on a runtime-adjustable parameter and the bathymetry.
The deeper the water column, the more GL points are used.