Make posting of EKE for visualization optional
EKE was being posted to the database at every timestep to demo
the ability to stream data out of the model and visualize it in
python. However, for long simulations, this caused the Redis
database nodes to run out of memory. A new runtime parameter
is thus introduced to avoid this.